app/pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["Maksat Hemdemow <thehighestintheroom00@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.101.0"
uvicorn = "^0.23.2"
sqlalchemy = "^2.0.19"
flake8 = "^6.1.0"
black = "^23.7.0"
alembic = "^1.11.2"
python-jose = "^3.3.0"
python-decouple = "^3.8"
passlib = "^1.7.4"
asyncpg-utils = "^0.6.0"
asyncpg = "^0.28.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
